Move TestServer and HTTPTestServer into test dir
